tests/transient: do use constructed `txdata` + other conveniences (test still fails).
No longer using a hard-coded list of recipients, but instead
generating a variable-sized one, based on `nrecipients` variable.
(Should help with benchmarking later on.)

There seems to be an issue with the program itself, though: the
balance assertion at the end of the test fails. I.e., the
recipient account's balance does not increase.