transient: for now, revert if a transfer fails.
This was actually a change that wasn't committed - coming back to
this after a hardware/wetware failure, so somewhat of a guesswork.

...

GOOD NEWS EVERYONE!

The reason why this approach doesn't work - after tracing - turns
out to be a "bug".

When the transient code is running, the "transaction data" is
all available as _code_ (doh!..), not _data_. So, `CODECOPY`
should have been used instead of `CALLDATALOAD`. Either that, or
some other method of copying portions of currently-running code to
memory/stack.

Will try to fix in subsequent commits.