transient: use `codecopy` instead of `calldataload`.
As explained in previous commit message, call data will be empty
at the bottom of the execution stack, as all transaction data is
present in this frame as code.

After this commit, there's also no longer need to maintain strict
alignment in transaction data, as use of the `(bytecodesize)` macro
provides that. It wasn't used, because I forgot it's available.

There is _still_ some issue in the code: recipient balances are
not increased.